work_stream/requirements.txt
fastapi
langchain
langchain-openai
langgraph
python-dotenv
requests
uvicorn
pydantic